projects/simple-langchain-langgraph-agent/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
python-dotenv>=1.0.1
requests>=2.32.0